Move component to own directory
