+++ context/app/static/js/components/Search/SelectedFilter/index.js
@@ -1,0 +1,3 @@
+import SelectedFilter from './SelectedFilter';
+
+export default SelectedFilter;
